revisi

+++ database/migrations/2025_12_04_141044_make_jumlah_buka_faktur_nullable_in_submissions.php
@@ -12,10 +12,13 @@ return new class extends Migration
     public function up()
     {
         Schema::table('submissions', function (Blueprint $table) {
-            $table->integer('jumlah_buka_faktur')->nullable()->change();
+            $table->bigInteger('jumlah_buka_faktur')->nullable()->change();
         });
     }
 
+    /**
+     * Reverse the migrations.
+     */
     public function down()
     {
         Schema::table('submissions', function (Blueprint $table) {
